refactor: station 패키지 이동
